refactor: changing vite ts file to js
