+++ extensions/kubectl-cli/vitest.config.js
