+++ extensions/registries/vitest.config.js
