+++ package.json
@@ -36,7 +36,8 @@
     "compile:pull-request": "cross-env MODE=production pnpm run build && electron-builder build --publish never --config .electron-builder.config.cjs",
     "compile:current": "cross-env MODE=production pnpm run build && electron-builder build --config .electron-builder.config.cjs",
     "test": "pnpm run test:unit && pnpm run test:e2e",
-    "test:unit": "pnpm run test:main && pnpm run test:preload && pnpm run test:preload-docker-extension && pnpm run test:preload-webview && pnpm run test:ui && pnpm run test:renderer && pnpm run test:scripts:stylesheet && pnpm run test:tools && pnpm run test:extensions && pnpm run test:website",
+    "test:unit": "vitest --run --config=vitest.config.unit",
+    "test:unit:coverage": "pnpm test:unit --coverage",
     "test:e2e": "pnpm run test:e2e:build && pnpm run test:e2e:run",
     "test:e2e:build": "cross-env NODE_ENV=development MODE=development DEBUG=pw:browser pnpm run build",
     "test:e2e:run": "xvfb-maybe --auto-servernum --server-args='-screen 0 1280x960x24' -- npx playwright test tests/playwright/src/specs/ --grep-invert @k8s_e2e",
@@ -54,26 +55,25 @@
     "test:e2e:remote:run": "xvfb-maybe --auto-servernum --server-args='-screen 0 1280x960x24' -- npx playwright test tests/playwright/src/special-specs/podman-remote/ -g @podman-remote",
     "test:e2e:update": "pnpm run test:e2e:build && pnpm run test:e2e:update:run",
     "test:e2e:update:run": "xvfb-maybe --auto-servernum --server-args='-screen 0 1280x960x24' -- npx playwright test tests/playwright/src/special-specs/installation/ -g @update-install",
-    "test:main": "vitest --run --project=main --passWithNoTests --coverage",
-    "test:preload": "vitest --run --project=preload --passWithNoTests --coverage",
-    "test:preload-docker-extension": "vitest --run --project=preload-docker-extension --passWithNoTests --coverage",
-    "test:preload-webview": "vitest --run --project=preload-webview --coverage",
-    "test:extensions": "vitest --run --project=extensions --passWithNoTests --coverage && pnpm run test:extensions:compose && pnpm run test:extensions:kind && pnpm run test:extensions:docker && pnpm run test:extensions:lima && pnpm run test:extensions:kube && pnpm run test:extensions:podman && pnpm run test:extensions:registries && pnpm run test:extensions:kubectl-cli",
-    "test:extensions:kind": "vitest --run --project=kind --passWithNoTests --coverage",
-    "test:extensions:compose": "vitest --run --project=compose --passWithNoTests --coverage",
-    "test:extensions:docker": "vitest --run --project=docker --passWithNoTests --coverage ",
-    "test:extensions:kube": "vitest --run --project=kube-context --passWithNoTests --coverage ",
-    "test:extensions:lima": "vitest --run --project=lima --passWithNoTests --coverage ",
-    "test:extensions:podman": "vitest --run --project=podman --passWithNoTests --coverage ",
-    "test:extensions:registries": "vitest --run --project=registries --passWithNoTests --coverage ",
-    "test:extensions:kubectl-cli": "vitest --run --project=kubectl-cli --passWithNoTests --coverage ",
-    "test:renderer": "pnpm run build:preload:types && pnpm run build:ui && vitest --run --project=renderer --passWithNoTests --coverage",
-    "test:ui": "vitest --run --project=@podman-desktop/ui-svelte --passWithNoTests --coverage",
-    "test:tools": "vitest run tools/src --passWithNoTests --coverage",
-    "test:scripts:stylesheet": "vitest run ./scripts/generate-stylesheet.spec.ts --passWithNoTests",
+    "test:main": "vitest --run --config=vitest.config.unit --project=main --passWithNoTests",
+    "test:preload": "vitest --run --config=vitest.config.unit --project=preload --passWithNoTests",
+    "test:preload-docker-extension": "vitest --run --config=vitest.config.unit --project=preload-docker-extension --passWithNoTests",
+    "test:preload-webview": "vitest --run --config=vitest.config.unit --project=preload-webview",
+    "test:extensions": "vitest --run --config=vitest.config.unit --project=extensions --passWithNoTests && pnpm run test:extensions:compose && pnpm run test:extensions:kind && pnpm run test:extensions:docker && pnpm run test:extensions:lima && pnpm run test:extensions:kube && pnpm run test:extensions:podman && pnpm run test:extensions:registries && pnpm run test:extensions:kubectl-cli",
+    "test:extensions:kind": "vitest --run --config=vitest.config.unit --project=kind --passWithNoTests",
+    "test:extensions:compose": "vitest --run --config=vitest.config.unit --project=compose --passWithNoTests",
+    "test:extensions:docker": "vitest --run --config=vitest.config.unit --project=docker --passWithNoTests",
+    "test:extensions:kube": "vitest --run --config=vitest.config.unit --project=kube-context --passWithNoTests ",
+    "test:extensions:lima": "vitest --run --config=vitest.config.unit --project=lima --passWithNoTests ",
+    "test:extensions:podman": "vitest --run --config=vitest.config.unit --project=podman --passWithNoTests ",
+    "test:extensions:registries": "vitest --run --config=vitest.config.unit --project=registries --passWithNoTests ",
+    "test:extensions:kubectl-cli": "vitest --run --config=vitest.config.unit --project=kubectl-cli --passWithNoTests",
+    "test:renderer": "pnpm run build:preload:types && pnpm run build:ui && vitest --run --project=renderer --passWithNoTests",
+    "test:ui": "vitest --run --config=vitest.config.unit --project=@podman-desktop/ui-svelte --passWithNoTests",
+    "test:tools": "vitest --run --config=vitest.config.unit --project=release-notes-generator --passWithNoTests",
+    "test:scripts:stylesheet": "vitest --run --config=vitest.config.unit --project=scripts --passWithNoTests",
     "test:watch": "vitest watch",
-    "test:website": "vitest run -r website --passWithNoTests --coverage",
-    "test:unit:coverage": "vitest --run --config=vitest.config.unit.ts --coverage",
+    "test:website": "vitest --run --config=vitest.config.unit --project=docs --passWithNoTests",
     "watch": "node scripts/watch.mjs",
     "format:check": "biome format && prettier --cache --check \"website/**/*.md\"",
     "format:clean": "rimraf node_modules/.cache/prettier",
